+++ docs/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 # Version 4.x
 
+## v4.2.0.0 (July 5, 2023)
+
+* `INFO` Minor tweaks and stuff
+
 ## v4.0.0 (March 8, 2023)
 
 * `INFO` Initial v4.x release